+++ aspnetcore/migration/20_21.md
@@ -113,7 +113,7 @@ The preceding code changes are detailed in:
 * [GDPR support in ASP.NET Core](xref:security/gdpr) for `CookiePolicyOptions` and `UseCookiePolicy`.
 * [HTTP Strict Transport Security Protocol (HSTS)](xref:security/enforcing-ssl#http-strict-transport-security-protocol-hsts) for `UseHsts`.
 * [Require HTTPS](xref:security/enforcing-ssl#require-https) for `UseHttpsRedirection`.
-* [SetCompatibilityVersion](xref:fundamentals/startup#setcompatibilityversion) for `SetCompatibilityVersion(CompatibilityVersion.Version_2_1)`.
+* [SetCompatibilityVersion](xref:mvc/compatibility-version) for `SetCompatibilityVersion(CompatibilityVersion.Version_2_1)`.
 
 ## Changes to authentication code
 
@@ -158,7 +158,7 @@ This section outlines the steps to replace the ASP.NET Core 2.0 template generat
   * *Extensions*
 * Build the project.
 * [Scaffold Identity](xref:security/authentication/scaffold-identity) into your project:
-  * Select the projects exiting *_Layout.cshtml* file.
+  * Select the projects exiting *\_Layout.cshtml* file.
   * Select the **+** icon on the right side of the **Data context class**. Accept the default name.
   * Select **Add** to create a new Data context class. Creating a new data context is required for to scaffold. You remove the new data context in the next section.
 
@@ -166,12 +166,12 @@ This section outlines the steps to replace the ASP.NET Core 2.0 template generat
 
 * Delete the Identity scaffolder generated `IdentityDbContext` derived class in the *Areas/Identity/Data/* folder.
 * Delete *Areas/Identity/IdentityHostingStartup.cs*
-* Update the *_LoginPartial.cshtml* file:
-  * Move *Pages/_LoginPartial.cshtml* to *Pages/Shared/_LoginPartial.cshtml*
+* Update the *\_LoginPartial.cshtml* file:
+  * Move *Pages/\_LoginPartial.cshtml* to *Pages/Shared/\_LoginPartial.cshtml*
   * Add `asp-area="Identity"` to the form and anchor links.
   * Update the `<form />` element to `<form asp-area="Identity" asp-page="/Account/Logout" asp-route-returnUrl="@Url.Page("/Index", new { area = "" })" method="post" id="logoutForm" class="navbar-right">`
 
-  The following code shows the updated *_LoginPartial.cshtml* file:
+  The following code shows the updated *\_LoginPartial.cshtml* file:
 
   [!code-cshtml[Main](20_21/sample/_LoginPartial.cshtml?highlight=8,11,22,23)]
 
@@ -183,15 +183,15 @@ Update `ConfigureServices` with the following code:
 
 ### The layout file
 
-* Move *Pages/_Layout.cshtml* to *Pages/Shared/_Layout.cshtml*
+* Move *Pages/\_Layout.cshtml* to *Pages/Shared/\_Layout.cshtml*
 * The *Layout.cshtml* file has the following changes:
 
   * `<partial name="_CookieConsentPartial" />` is added. For more information, see [GDPR support in ASP.NET Core](xref:security/gdpr).
   * jQuery changes from 2.2.0 to 3.3.1
 
-### _ValidationScriptsPartial.cshtml
+### \_ValidationScriptsPartial.cshtml
 
-* *Pages/_ValidationScriptsPartial.cshtml* moves to  *Pages/Shared/_ValidationScriptsPartial.cshtml*
+* *Pages/\_ValidationScriptsPartial.cshtml* moves to *Pages/Shared/\_ValidationScriptsPartial.cshtml*
 * *jquery.validate/1.14.0* changes to *jquery.validate/1.17.0*
 
 ### New files
@@ -212,7 +212,7 @@ The *Layout.cshtml* file has the following changes:
 * `<partial name="_CookieConsentPartial" />` is added.
 * jQuery changes from 2.2.0 to 3.3.1
 
-### _ValidationScriptsPartial.cshtml
+### \_ValidationScriptsPartial.cshtml
 
 *jquery.validate/1.14.0* changes to *jquery.validate/1.17.0*
 
@@ -228,5 +228,5 @@ See [GDPR support in ASP.NET Core](xref:security/gdpr) for information on the pr
 ## Additional changes
 
 * If hosting the app on Windows with IIS, install the latest [.NET Core Hosting Bundle](xref:host-and-deploy/iis/index#install-the-net-core-hosting-bundle).
-* [SetCompatibilityVersion](xref:fundamentals/startup#setcompatibilityversion)
+* [SetCompatibilityVersion](xref:mvc/compatibility-version)
 * [Transport configuration](xref:fundamentals/servers/kestrel#transport-configuration)
