+++ aspnetcore/mvc/compatibility-version/samples/2.x/CompatibilityVersionSample/CompatibilityVersionSample.csproj
@@ -5,11 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Folder Include="wwwroot\" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App"  />
+    <PackageReference Include="Microsoft.AspNetCore.App"  Version="2.1.2" />
   </ItemGroup>
 
 </Project>
