+++ aspnetcore/mvc/compatibility-version/samples/2.x/CompatibilityVersionSample/Program.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 
-namespace WebApp1
+namespace CompatibilityVersionSample
 {
     public class Program
     {
